requirements.txt
Django==1.9
django-crispy-forms==1.6.0
django-friendship==1.3.3
django-registration-redux==1.4
Pillow==3.3.0
